requirements.txt
arrow==0.10.0
binaryornot==0.4.0
chardet==2.3.0
click==6.6
cookiecutter==1.5.0
Django==1.10.5
django-appconf==1.0.2
django-bootstrap3==8.0.0
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-easy-maps==0.9.3
django-extensions==1.7.5
djangorestframework==3.5.3
Faker==0.7.7
Flask==0.12
future==0.16.0
geopy==1.11.0
itsdangerous==0.24
Jinja2==2.8.1
jinja2-time==0.2.0
MarkupSafe==0.23
mysql-connector==2.1.4
olefile==0.44
pbr==1.10.0
Pillow==4.0.0
pkg-resources==0.0.0
poyo==0.4.0
pygame==1.9.2
python-dateutil==2.6.0
six==1.10.0
South==1.0.2
sqlparse==0.2.2
stevedore==1.19.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.11.15
whichcraft==0.4.0
